test(logic): split research_phase_test under 400-line ceiling (Refs #2216)
The PR-incremental repo.logic_test_file_size check failed because
research_phase_test.dart was 568 physical lines after fixture extraction.
Split tests by concern to bring each file under 400 lines:

- research_phase_test.dart -> direct resolveResearchPhase tests (155 lines)
- research_phase_funding_test.dart -> turn-resolution funding/prereq tests (273 lines)
- research_phase_tech_effects_test.dart -> University/Money Lending/Banking
  and duplicate-slot tests (160 lines)

Drop unused colonizethis_data import from research_phase_test_support.dart.
All 16 original test cases preserved.
