+++ packages/colonizethis_logic/test/research_phase_test_support.dart
@@ -1,4 +1,3 @@
-import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 
 /// Shared fixtures for [research_phase_test.dart].
